requirements.txt
python-telegram-bot==21.3
SQLAlchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
greenlet==3.0.3
